pyproject.toml
[project]
name = "jetson-orin-nano-oled"
version = "0.1.0"
description = "Test SPI and I2C connections to OLED displays from NVIDIA Jetson Orin Nano"
readme = "README.md"
authors = [
    { name = "Lachlan Hurst", email = "lachlan.hurst@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "jetson-gpio>=2.1.9",
    "luma-oled>=3.14.0",
]

[project.scripts]
oled-i2c = "jetson_orin_nano_oled.oled_i2c:main"
oled-spi = "jetson_orin_nano_oled.oled_spi:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
